PAS: Add error checking in recovery process
When PCL file API is used to write data to files, the files for different
users are located in subfolders <appId>/user/<user_no>/seat/<seat_no>/.
If the PCL file API is not called to write data to file, the folder
<appId>/user/<user_no>/seat/<seat_no>/ will not exist and backup will
not create the folder. In this case recovery will fail as there is no
error checking to check if source folder exists during recovery phase.

Similar error case is seen if /Data/mnt-c/<appId>/node folder is
not created.

This commit adds error checking in recovery and backup cases.

Change-Id: I2f9a01a89c567cf70a178c8f6142f714244311b0
Signed-off-by: Vignesh Raman <Vignesh_Raman@mentor.com>