Replace unknown oip-cpi-mandatory.target in systemd service file
With a well known systemd target. oip-cpi-mandatory.target is Continental
specific?

Do not know for sure that multi-user.target is the best one. Perhaps the
administrator service should be changed to be D-Bus activated in the future
since it is not required to be running all the time (if I have understood
things correctly).

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>